requirements.txt
## Primary Dependencies
Django==4.2.5
django-environ==0.11.2
djangorestframework==3.14.0
nltk==3.8.1
numpy==1.25.2
psycopg2==2.9.7
pyensign==0.9b0
scikit-learn==1.3.0

## Secondary Dependencies
asgiref==3.7.2
click==8.1.7
grpcio==1.57.0
joblib==1.3.2
protobuf==4.24.2
PyJWT==2.8.0
python-ulid==1.1.0
pytz==2023.3
regex==2023.8.8
scipy==1.11.2
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.66.1

## Setup Tools and Python
pip==23.2.1
setuptools==65.5.0